+++ src/entities/stub-template/index.ts
@@ -1,0 +1,4 @@
+export type { StubTemplate, StubTemplateInput, StubTemplateCategory } from "@/entities/stub-template/model/types";
+export { stubTemplateCategoryValues } from "@/entities/stub-template/model/types";
+export { builtInStubTemplates } from "@/entities/stub-template/model/builtin-templates";
+export { useStubTemplates } from "@/entities/stub-template/model/use-stub-templates";