+++ src/features/create-stub-mapping-from-template/index.ts
@@ -1,0 +1,2 @@
+export type { CreateStubMappingFromTemplateNavigationState } from "@/features/create-stub-mapping-from-template/model/use-create-stub-mapping-from-template";
+export { useCreateStubMappingFromTemplate } from "@/features/create-stub-mapping-from-template/model/use-create-stub-mapping-from-template";